# chacha20: improve generic implementation performance